[vtadmin-web] Move single-entity view components into subfolders
