+++ web/vtadmin/src/components/App.tsx
@@ -25,10 +25,10 @@ import { Clusters } from './routes/Clusters';
 import { Gates } from './routes/Gates';
 import { Keyspaces } from './routes/Keyspaces';
 import { Schemas } from './routes/Schemas';
-import { Schema } from './routes/Schema';
-import { Stream } from './routes/Stream';
+import { Schema } from './routes/schema/Schema';
+import { Stream } from './routes/stream/Stream';
 import { Workflows } from './routes/Workflows';
-import { Workflow } from './routes/Workflow';
+import { Workflow } from './routes/workflow/Workflow';
 import { VTExplain } from './routes/VTExplain';
 import { Keyspace } from './routes/keyspace/Keyspace';
 import { Tablet } from './routes/tablet/Tablet';
