+++ web/vtadmin/src/components/routes/schema/Schema.tsx
@@ -17,16 +17,16 @@ import * as React from 'react';
 import { Link, useParams } from 'react-router-dom';
 
 import style from './Schema.module.scss';
-import { useSchema, useVSchema } from '../../hooks/api';
-import { Code } from '../Code';
-import { useDocumentTitle } from '../../hooks/useDocumentTitle';
-import { getVindexesForTable } from '../../util/vschemas';
-import { ContentContainer } from '../layout/ContentContainer';
-import { NavCrumbs } from '../layout/NavCrumbs';
-import { WorkspaceHeader } from '../layout/WorkspaceHeader';
-import { WorkspaceTitle } from '../layout/WorkspaceTitle';
-import { Tooltip } from '../tooltip/Tooltip';
-import { KeyspaceLink } from '../links/KeyspaceLink';
+import { useSchema, useVSchema } from '../../../hooks/api';
+import { Code } from '../../Code';
+import { useDocumentTitle } from '../../../hooks/useDocumentTitle';
+import { getVindexesForTable } from '../../../util/vschemas';
+import { ContentContainer } from '../../layout/ContentContainer';
+import { NavCrumbs } from '../../layout/NavCrumbs';
+import { WorkspaceHeader } from '../../layout/WorkspaceHeader';
+import { WorkspaceTitle } from '../../layout/WorkspaceTitle';
+import { Tooltip } from '../../tooltip/Tooltip';
+import { KeyspaceLink } from '../../links/KeyspaceLink';
 
 interface RouteParams {
     clusterID: string;
